requirements.txt
# 麻将听牌分析器 v3.0 依赖包
Pillow>=8.0.0              # 图片处理库
requests>=2.25.0           # HTTP请求库（用于下载素材）